+++ Sources/Core/Application Services/ConfigureMainQueueApplicationService.swift
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.swift
+//  AppwiseCore
+//
+//  Created by David Jennes on 17/09/16.
+//  Copyright © 2016 Appwise. All rights reserved.
+//
+
+import Foundation
+
+final class ConfigureMainQueueApplicationService: NSObject, ApplicationService {
+	func application(_ application: UIApplication, willFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey : Any]? = nil) -> Bool {
+		DispatchQueue.configureMainQueue()
+
+		return true
+	}
+}